Website docs style patch (#11286)
This PR is in response to https://github.com/fleetdm/fleet/issues/10420,
to improve the readability of the docs. In particular,
[docs/using-fleet/mdm-setup](https://fleetdm.com/docs/using-fleet/mdm-setup).

**CSS changes**
- Adjusted margins and paddings on `<ol>` and `<ul>` elements. As well
as h-tags and code blocks.
- Added styling to docs `<h3>` tags to help differentiate sub-sections.
This helps to make the docs pages easier to scan.
- Reduced the boldness of `<b>` and `<strong>` tags. The recently
updated font, Inter, renders bold a little too heavily. (This is a
site-wide change.)
- Changed the default font color to `@core-fleet-black-75` and made sure
that h-tags are set to `@core-fleet-black`. The softer contrast helps
reduce fatigue while reading large blocks of text (while still
conforming to recommended contrast levels). This site-wide change brings
the text styling in sync with Figma.

**Content changes**
- Reformatted content on
[docs/using-fleet/mdm-setup](https://fleetdm.com/docs/using-fleet/mdm-setup)
to fix Markdown issues that were causing sections to render incorrectly,
and were also breaking the "On this page" side nav.
- Made grammar and content fixes to improve readability and flow.

# Checklist for submitter

If some of the following don't apply, delete the relevant line.

- [x] Manual QA for all new/changed functionality

---------

Co-authored-by: Eric <eashaw@sailsjs.com>