Update REST-API.md
Current REST API documentation for activities is outdated. Since we have automatically generated documentation for activity types on the Audit activities page, with all the necessary details I think it's a good idea to link REST API docs to this one.

REST API docs will have a description of what will be returned with example and specific information about each activity type can be found on the Audit activities page, which is always up-to-date.