Update ceo-handbook.md (#12280)
Proposed changes:
- I spotted a typo
- I updated references to previous staff members while reading.
- I spelled out "Executive Assistant" in the first couple of instances
and included "(EA)." It may seem an obvious acronym, but I'm a dumbo,
and the first time I heard EA mentioned, I thought we were talking about
Electronic Arts, the software developer. That is not entirely true, but
you get my point 😉

There are also two more references to @mentioning Charlie in Slack on
lines 148 and 150 that I don't know who to update to, so I have left
them untouched for the time being.