feat(terraform): add support for volumes and mount points in fleet_config
This commit adds support for volumes and mount points in the fleet_config
variable for the byo-vpc/byo-db/byo-ecs Terraform module. The volumes and mount
points are defined as lists of maps in the fleet_config variable. The maps
contain the following keys: name, efs_volume_configuration, docker_volume_configuration, and host_path.

The commit also adds the depends_on key to the fleet_config variable. This key
is used to specify dependencies between resources in the module. Finally, the
commit updates the variables.tf files in the byo-vpc, byo-db, and byo-ecs
directories to include the new keys in the fleet_config variable.