Handbook: Rename emergency drafting to expedited drafting (#12433)
It seems like we’ve been incorporating “emergency drafting” more
frequently into sprint cycles. This totally makes sense given that we
want to be an agile, design-driven company. It might be an easy cultural
win to instead call it something like “expedited drafting” to soften the
dissonance that comes from the sense of being in a constant state of
emergency.