Update product groups handbook (#12934)
Updating to reflect the temporary (1 sprint) move of Martin to CX. I
will update again next sprint.

It also seems like we can remove the old development groups language and
call out. What do you think?

I'm also removing the `*` next to names because we weren't indicating
what that meant, which seems unnecessary. It is up to the EM to
understand the capacity of their team.
.