Update product group assignees (#13704)
The product group capacities is # of engineers * 13. #g-website seemed
way too high given it's just Eric, so I dropped that to 13. Note that
he's also being counted for 13 on the CX capacity, so we're
double-counting Eric.

His vuln dashboard work is tracked on the CX board, which is why we're
counting him there.