Feature MS-MDM management endpoint logic (#14293)
This relates  to #12841 

The code implements the logic to track both stateless and stateful
commands.

Stateless commands are the commands sent by fleetctl/UI and enqueued
into the mdm pending commands table. These commands just contains
actions to be performed on a given deviceID. The don't yet have the
MessageID, SessionID and CmdId information as these data gets populated
during the response to a given MDM management request.

On the other hand, Stateful commands are the commands that contain the
MessageID, SessionID and CmdId and that were sent to the device. These
commands get stored on the commands table for tracking purposes.

The flow looks like this:

- Fleetctl calls into Fleet server API to send a new command to a
device.
- The API ends up receiving this data and "enqueuing" this command to
the pending commands table through the MDMWindowsInsertPendingCommand
function. At this point, these command request are stateless as these
don't contain MessageID, SessionID, and CmdID information
- Whenever a new management request coming from the MDM client stack
(device) hits the MDE2ManagementPath URL (mdmMicrosoftManagementEndpoint
handler), the pending cmds for the calling deviceID are retrieved and
converted into SynCML response commands with actual state data. These
commands are stored on the commands table for tracking purposes.
- The SyncML command stored on the commands table contains state
information and represents an actual command sent to the device.
- The device will reply to these commands at some point by calling again
into the management URL. The response for these commands contains an
error code that indicates the execution state for a given command. The
response are also SyncML commands with Status/Results protocol verbs.
These error code are tracked on the commands table. The entry point to
know to which command a given response belongs is the state information
contained in the MessageID, SessionID, and CmdID fields.

If some of the following don't apply, delete the relevant line.

- [X] Changes file added for user-visible changes in `changes/` or
`orbit/changes/`.
See [Changes
files](https://fleetdm.com/docs/contributing/committing-changes#changes-files)
for more information.
- [X] Added/updated tests
- [X] Manual QA for all new/changed functionality