Show kickstart flag warning only when set (#13924)
The initial implementation of this warning in #12072 used the same `if
!c.Bool("disable-kickstart-softwareupdated")` check as the old code, but
the body of the `if` was the kickstart which was being skipped, now it
is the warning. So currently the warning is showing only when the flag
is *not* used.

I'm not building this software myself, just an end user who had a
version of this assigned to my machine. Since I noticed the warning in
logs and figured I'd submit a PR, please feel free to take it over to
finish the checklist.