Update why-fleet.md
I moved the "What's it for?" heading.  I realized after checking the new content on the site that placing it between the new paragraphs disrupted the message.

I also addressed some grammar and formatting issues. Removed some redundant words, and rephrased a couple of sentences to give a more personable tone.