Further clarify observer permissions to view queries & query reports (#15043)
It's a bit unclear from the permissions docs _how_ observers are able to
view all queries & their reports, leading to bug reports like
[this](https://github.com/fleetdm/fleet/issues/15009).
I think the intended behavior is that observers can view all queries and
their reports **via the API** but not in the UI or fleetctl. Updated
notes to clarify. (Let me know if my interpretation is incorrect!)