Standardize where to specify design changes in the bug report template  (#15027)
Since a lot of bugs end up needing additional product design work, I
propose adding a (commented-out by default) section to this template to
standardize where to add design changes, once settled.

Reasoning: in estimation sessions, it can sometimes be hard to find this
information: sometimes it's in the comments, sometimes it's been added
to the description... either way, its not always obvious to spot. I
think it will help us move quicker if there's a consistent heading to
look for.

(Also, open to suggestions for other ways of wording that heading! This
is just the way I've been adding it to issue descriptions lately.)