Add performance feedback
Mention of performance feedback needs to remain on the communications page, to be clear of the expectation for all Fleeties. It cannot live solely on the Leadership page (which has a different target audience). Reverting a change that happened recently, and linking to the manager guidance in the Leadership page.