Update standard-query-library.yml
It is very easy for data collection like this to veer into double / triple negative mulitverse of madness stuff...

That said, I may have a lack of understanding about how the product works, i.e., that a query literally must return a 0 value & not null in order to "pass" in a policy. If so, then this works as expected.

However, if a query just needs to return empty (null) & 0 is implied in the logic that sets a policy flag to green or red, then, as a rule, queries like this should be always be simplified & should default to using "positive" as opposed to "negative" logic, i.e., check if something exists, never check if a thing does NOT exist.