Signing authority
Updated the section on getting a customer agreement signed and bifurcated responsibility between the CEO for customer contracts and Head of BizOps for vendors/suppliers/service providers.  

Mike had suggested me as the signing authority, but since Joanne and I communicate daily I felt it would be easy to maintain a degree of separation between the reviewing and signing authority and the benefit of an extra set of eyes asking if it makes sense to sign up a new vendor while still allowing for prompt execution of contracts with vendors because of the proximity of me as the reviewer to Joanne as the signatory and would prevent miscommunications.  

This change came out of a convo with Mike at the start of the month documented in this issue: https://app.zenhub.com/workspaces/-g-business-operations-63f3dc3cc931f6247fcf55a9/issues/gh/fleetdm/classified/1085 

Mike suggested a second docusign account for supplier contracts and I was thinking getting one for joanne@fleet would be appropriate.  Once we have that I will update this entry again with her docusign info.