Add wipe activity and improve host past activites rendering (#16957)
relates to #10489

Added the Wipe action to the global action feed.

Also refactor of the Host past activities to render from a configuration
instead of switch conditions. We've created a generic HostActivityItem
component that renders all the common UI. The individual activities use
this as a wrapper component and then render their own details. We also
improve on the typing for past activities.

This was done to make it simpler to create new actions and prevent a
huge file like we have for the dashboard actions
`DashboardPage/cards/ActivityItems.tsx` which will help with
maintainability.

- [x] Manual QA for all new/changed functionality