Fix wrong number of host in mdm solutions modal and fix mdm solutions table UI to work for null named solutions (#17336)
relates to #16837, #17334, #17335

This fixes a UI bug for the case where the mdm solution name can be
null. We now handle this case properly and show the mdm solution data in
the modal.

This also fixes a UI bug where we showed the incorrect number of hosts
in the mdm solutions modal.

There is various cleanup here to the js and scss code in this PR too. 

- [x] Added/updated tests
- [x] Manual QA for all new/changed functionality