Support for Terraforming FleetDM Teams
This project adds support for terraforming teams in FleetDM. If you
have 100+ teams, managing them is is prone to error, mistakes, lost
configuration, and general pain. An industry standard tool like
terraform can unify this configuration as code.

In order to do this, I wrote a terraform provider that on one end
talks to the FleetDM api, and on the other end implements an interface
for terraform. More information is in the README.

A small sample `main.tf` file is supplied.