🔐 feat(security): add support for sts:ExternalId in AWS STS AssumeRole calls
add support for sts:ExternalId in AWS STS AssumeRole calls. This is a security feature that helps protect against the confused deputy problem. It is an optional unique identifier that can be used by the principal (Fleet) assuming the role to assert its identity.

📝 docs(fleet-server-configuration): add documentation for sts_external_id for AWS services

add documentation for the sts_external_id configuration option for Firehose, Kinesis, Lambda, SES, S3, and Packaging S3. This option is used for AWS STS External ID authentication, typically in conjunction with an STS role ARN to ensure that only the intended AWS account can assume the role.