Update features.yml: iOS/iPadOS dates (#19341)
- iOS/iPadOS launch pushed from 6/30 => 7/15
- Why? We didn't have enough design nor engineering capacity. Bugs,
dogfooding feature requests and features for the "Improve the
self-service tech eval experience" were prioritized first

Noah: I don't think we'll be ready for the public launch until we ship
the following:
- #19319 

Why? We want iOS/iPadOS support to feel first class when we publicly
announce it.

Note that Fleet is shipping enrollment + custom MDM commands (#18119)
for iOS/iPadOS and dogfooding profiles (#18866) starting 2024-06-03.
`customer-eponym` will be an early adopter.