Tf software installers fix (#19715)
Fixed a typo in the name of the software installers bucket prefix env
var in terraform modules and removed the default prefix as the code
seems to create this prefix automatically. An additional prefix is still
supported for configuration, but can be left empty and will be so as the
default.